Fix #17 (Uncaught TypeError: Unsupported operand types: string * int)
Multiplication of an `int` and non-numeric strings (in this case empty
strings) is no longer supported as of PHP 8.0.0.  We fix this by
explicit casting.

It should be noted that the way `$month_input` and `$year_input` are
constructed is erroneous.  Preferable, we should either expect a
`$_POST` *or* `$_GET` parameter, but we definitely should not
concatenate, in case both are provided.  Still, I don't dare to fix
this (at least right now), and it is a separate issue anyway.